+++ packages/contracts/contracts/core/IDAO.sol
@@ -5,7 +5,7 @@ pragma solidity 0.8.17;
 /// @title IDAO
 /// @author Aragon Association - 2022-2023
 /// @notice The interface required for DAOs within the Aragon App DAO framework.
-abstract contract IDAO {
+interface IDAO {
     struct Action {
         address to; // Address to call
         uint256 value; // Value to be sent with the call (for example ETH if on mainnet)
@@ -23,11 +23,11 @@ abstract contract IDAO {
         address _who,
         bytes32 _permissionId,
         bytes memory _data
-    ) external view virtual returns (bool);
+    ) external view returns (bool);
 
     /// @notice Updates the DAO metadata (e.g., an IPFS hash).
     /// @param _metadata The IPFS hash of the new metadata object.
-    function setMetadata(bytes calldata _metadata) external virtual;
+    function setMetadata(bytes calldata _metadata) external;
 
     /// @notice Emitted when the DAO metadata is updated.
     /// @param metadata The IPFS hash of the new metadata object.
@@ -44,7 +44,7 @@ abstract contract IDAO {
         bytes32 callId,
         Action[] memory _actions,
         uint256 _allowFailureMap
-    ) external virtual returns (bytes[] memory, uint256);
+    ) external returns (bytes[] memory, uint256);
 
     /// @notice Emitted when a proposal is executed.
     /// @param actor The address of the caller.
@@ -76,11 +76,7 @@ abstract contract IDAO {
     /// @param _token The address of the token or address(0) in case of the native token.
     /// @param _amount The amount of tokens to deposit.
     /// @param _reference The reference describing the deposit reason.
-    function deposit(
-        address _token,
-        uint256 _amount,
-        string calldata _reference
-    ) external payable virtual;
+    function deposit(address _token, uint256 _amount, string calldata _reference) external payable;
 
     /// @notice Emitted when a token deposit has been made to the DAO.
     /// @param sender The address of the sender.
@@ -100,32 +96,13 @@ abstract contract IDAO {
     /// @param amount The amount of native tokens deposited.
     event NativeTokenDeposited(address sender, uint256 amount);
 
-    /// @notice Withdraw (native) tokens from the DAO with a withdraw reference string.
-    /// @param _token The address of the token and address(0) in case of the native token.
-    /// @param _to The target address to send (native) tokens to.
-    /// @param _amount The amount of (native) tokens to withdraw.
-    /// @param _reference The reference describing the withdrawal reason.
-    function withdraw(
-        address _token,
-        address _to,
-        uint256 _amount,
-        string memory _reference
-    ) external virtual;
-
-    /// @notice Emitted when a (native) token withdrawal has been made from the DAO.
-    /// @param token The address of the withdrawn token or address(0) in case of the native token.
-    /// @param to The address of the withdrawer.
-    /// @param amount The amount of tokens withdrawn.
-    /// @param _reference The reference describing the withdrawal reason.
-    event Withdrawn(address indexed token, address indexed to, uint256 amount, string _reference);
-
     /// @notice Setter for the trusted forwarder verifying the meta transaction.
     /// @param _trustedForwarder The trusted forwarder address.
-    function setTrustedForwarder(address _trustedForwarder) external virtual;
+    function setTrustedForwarder(address _trustedForwarder) external;
 
     /// @notice Getter for the trusted forwarder verifying the meta transaction.
     /// @return The trusted forwarder address.
-    function getTrustedForwarder() external virtual returns (address);
+    function getTrustedForwarder() external view returns (address);
 
     /// @notice Emitted when a new TrustedForwarder is set on the DAO.
     /// @param forwarder the new forwarder address.
@@ -133,7 +110,7 @@ abstract contract IDAO {
 
     /// @notice Setter for the [ERC-1271](https://eips.ethereum.org/EIPS/eip-1271) signature validator contract.
     /// @param _signatureValidator The address of the signature validator.
-    function setSignatureValidator(address _signatureValidator) external virtual;
+    function setSignatureValidator(address _signatureValidator) external;
 
     /// @notice Emitted when the signature validator address is updated.
     /// @param signatureValidator The address of the signature validator.
@@ -143,10 +120,7 @@ abstract contract IDAO {
     /// @param _hash The keccak256 hash of arbitrary length data signed on the behalf of `address(this)`.
     /// @param _signature Signature byte array associated with _data.
     /// @return magicValue Returns the `bytes4` magic value `0x1626ba7e` if the signature is valid.
-    function isValidSignature(bytes32 _hash, bytes memory _signature)
-        external
-        virtual
-        returns (bytes4);
+    function isValidSignature(bytes32 _hash, bytes memory _signature) external returns (bytes4);
 
     /// @notice Registers an ERC standard having a callback by registering its [ERC-165](https://eips.ethereum.org/EIPS/eip-165) interface ID and callback function signature.
     /// @param _interfaceId The ID of the interface.
@@ -156,5 +130,5 @@ abstract contract IDAO {
         bytes4 _interfaceId,
         bytes4 _callbackSelector,
         bytes4 _magicNumber
-    ) external virtual;
+    ) external;
 }